+++ path.py
@@ -62,6 +62,7 @@ import io
 import importlib
 import itertools
 import platform
+import ntpath
 
 try:
     import win32security
@@ -94,6 +95,7 @@ if PY2:
     text_type = __builtin__.unicode
     getcwdu = os.getcwdu
     map = itertools.imap
+    filter = itertools.ifilter
     FileNotFoundError = OSError
     itertools.filterfalse = itertools.ifilterfalse
 
@@ -173,6 +175,60 @@ class multimethod(object):
             functools.partial(self.func, owner) if instance is None
             else functools.partial(self.func, owner, instance)
         )
+
+
+class matchers(object):
+    # TODO: make this class a module
+
+    @staticmethod
+    def load(param):
+        """
+        If the supplied parameter is a string, assum it's a simple
+        pattern.
+        """
+        return (
+            matchers.Pattern(param) if isinstance(param, string_types)
+            else param if param is not None
+            else matchers.Null()
+        )
+
+    class Base(object):
+        pass
+
+    class Null(Base):
+        def __call__(self, path):
+            return True
+
+    class Pattern(Base):
+        def __init__(self, pattern):
+            self.pattern = pattern
+
+        def get_pattern(self, normcase):
+            try:
+                return self._pattern
+            except AttributeError:
+                pass
+            self._pattern = normcase(self.pattern)
+            return self._pattern
+
+        def __call__(self, path):
+            normcase = getattr(self, 'normcase', path.module.normcase)
+            pattern = self.get_pattern(normcase)
+            return fnmatch.fnmatchcase(normcase(path.name), pattern)
+
+    class CaseInsensitive(Pattern):
+        """
+        A Pattern with a ``'normcase'`` property, suitable for passing to
+        :meth:`listdir`, :meth:`dirs`, :meth:`files`, :meth:`walk`,
+        :meth:`walkdirs`, or :meth:`walkfiles` to match case-insensitive.
+
+        For example, to get all files ending in .py, .Py, .pY, or .PY in the
+        current directory::
+
+            from path import Path, matchers
+            Path('.').files(matchers.CaseInsensitive('*.py'))
+        """
+        normcase = staticmethod(ntpath.normcase)
 
 
 class Path(text_type):
@@ -537,7 +593,7 @@ class Path(text_type):
 
     # --- Listing, searching, walking, and matching
 
-    def listdir(self, pattern=None):
+    def listdir(self, match=None):
         """ D.listdir() -> List of items in this directory.
 
         Use :meth:`files` or :meth:`dirs` instead if you want a listing
@@ -545,18 +601,15 @@ class Path(text_type):
 
         The elements of the list are Path objects.
 
-        With the optional `pattern` argument, this only lists
-        items whose names match the given pattern.
+        With the optional `match` argument, a callable,
+        only return items whose names match the given pattern.
 
         .. seealso:: :meth:`files`, :meth:`dirs`
         """
-        if pattern is None:
-            pattern = '*'
-        return [
-            self / child
-            for child in os.listdir(self)
-            if self._next_class(child).fnmatch(pattern)
-        ]
+        match = matchers.load(match)
+        return list(filter(match, (
+            self / child for child in os.listdir(self)
+        )))
 
     def dirs(self, *args, **kwargs):
         """ D.dirs() -> List of this directory's subdirectories.
@@ -580,7 +633,7 @@ class Path(text_type):
 
         return [p for p in self.listdir(*args, **kwargs) if p.isfile()]
 
-    def walk(self, pattern=None, errors='strict'):
+    def walk(self, match=None, errors='strict'):
         """ D.walk() -> iterator over files and subdirs, recursively.
 
         The iterator yields Path objects naming each child item of
@@ -610,6 +663,8 @@ class Path(text_type):
             raise ValueError("invalid errors parameter")
         errors = vars(Handlers).get(errors, errors)
 
+        match = matchers.load(match)
+
         try:
             childList = self.listdir()
         except Exception:
@@ -620,7 +675,7 @@ class Path(text_type):
             return
 
         for child in childList:
-            if pattern is None or child.fnmatch(pattern):
+            if match(child):
                 yield child
             try:
                 isdir = child.isdir()
@@ -632,7 +687,7 @@ class Path(text_type):
                 isdir = False
 
             if isdir:
-                for item in child.walk(pattern, errors):
+                for item in child.walk(errors=errors, match=match):
                     yield item
 
     def walkdirs(self, *args, **kwargs):
@@ -1733,125 +1788,24 @@ def _permission_mask(mode):
     return functools.partial(op_map[op], mask)
 
 
-class CaseInsensitivePattern(text_type):
-    """
-    A string with a ``'normcase'`` property, suitable for passing to
-    :meth:`listdir`, :meth:`dirs`, :meth:`files`, :meth:`walk`,
-    :meth:`walkdirs`, or :meth:`walkfiles` to match case-insensitive.
-
-    For example, to get all files ending in .py, .Py, .pY, or .PY in the
-    current directory::
-
-        from path import Path, CaseInsensitivePattern as ci
-        Path('.').files(ci('*.py'))
-    """
-
-    @property
-    def normcase(self):
-        return __import__('ntpath').normcase
+class CaseInsensitivePattern(matchers.CaseInsensitive):
+    def __init__(self, value):
+        warnings.warn(
+            "Use matchers.CaseInsensitive instead",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        super(CaseInsensitivePattern, self).__init__(value)
 
 
 class FastPath(Path):
-    """
-    Performance optimized version of Path for use
-    on embedded platforms and other systems with limited
-    CPU. See #115 and #116 for background.
-    """
-
-    def listdir(self, pattern=None):
-        children = os.listdir(self)
-        if pattern is None:
-            return [self / child for child in children]
-
-        pattern, normcase = self.__prepare(pattern)
-        return [
-            self / child
-            for child in children
-            if self._next_class(child).__fnmatch(pattern, normcase)
-        ]
-
-    def walk(self, pattern=None, errors='strict'):
-        class Handlers:
-            def strict(msg):
-                raise
-
-            def warn(msg):
-                warnings.warn(msg, TreeWalkWarning)
-
-            def ignore(msg):
-                pass
-
-        if not callable(errors) and errors not in vars(Handlers):
-            raise ValueError("invalid errors parameter")
-        errors = vars(Handlers).get(errors, errors)
-
-        if pattern:
-            pattern, normcase = self.__prepare(pattern)
-        else:
-            normcase = None
-
-        return self.__walk(pattern, normcase, errors)
-
-    def __walk(self, pattern, normcase, errors):
-        """ Prepared version of walk """
-        try:
-            childList = self.listdir()
-        except Exception:
-            exc = sys.exc_info()[1]
-            tmpl = "Unable to list directory '%(self)s': %(exc)s"
-            msg = tmpl % locals()
-            errors(msg)
-            return
-
-        for child in childList:
-            if pattern is None or child.__fnmatch(pattern, normcase):
-                yield child
-            try:
-                isdir = child.isdir()
-            except Exception:
-                exc = sys.exc_info()[1]
-                tmpl = "Unable to access '%(child)s': %(exc)s"
-                msg = tmpl % locals()
-                errors(msg)
-                isdir = False
-
-            if isdir:
-                for item in child.__walk(pattern, normcase, errors):
-                    yield item
-
-    def __fnmatch(self, pattern, normcase):
-        """ Return ``True`` if `self.name` matches the given `pattern`,
-        prepared version.
-        `pattern` - A filename pattern with wildcards,
-            for example ``'*.py'``. The pattern is expected to be normcase'd
-            already.
-        `normcase` - A function used to normalize the pattern and
-            filename before matching.
-        .. seealso:: :func:`Path.fnmatch`
-        """
-        return fnmatch.fnmatchcase(normcase(self.name), pattern)
-
-    def __prepare(self, pattern, normcase=None):
-        """ Prepares a fmatch_pattern for use with ``FastPath.__fnmatch`.
-        `pattern` - A filename pattern with wildcards,
-            for example ``'*.py'``. If the pattern contains a `normcase`
-            attribute, it is applied to the name and path prior to comparison.
-        `normcase` - (optional) A function used to normalize the pattern and
-            filename before matching. Defaults to :meth:`self.module`,
-            which defaults to :meth:`os.path.normcase`.
-        .. seealso:: :func:`FastPath.__fnmatch`
-        """
-        if not normcase:
-            normcase = getattr(pattern, 'normcase', self.module.normcase)
-        pattern = normcase(pattern)
-        return pattern, normcase
-
-    def fnmatch(self, pattern, normcase=None):
-        if not pattern:
-            raise ValueError("No pattern provided")
-
-        pattern, normcase = self.__prepare(pattern, normcase)
-        return self.__fnmatch(pattern, normcase)
+    def __init__(self, *args, **kwargs):
+        warnings.warn(
+            "Use Path, as FastPath no longer holds any advantage",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        super(FastPath, self).__init__(*args, **kwargs)
 
 
 def patch_for_linux_python2():